requirements.txt
streamlit
python-dotenv
supabase
pandas
langchain-core
langchain-groq
langgraph
typing_extensions